packages/octane/src/edition/edition-manager.ts
import { Provider, Wallet } from "fuels";
import { NFTMetadata, Edition as EditionType } from "../common/types";

/**
 * @class EditionManager
 * @classdesc Manages editions within the Octane SDK on the Fuel network.
 */
export class EditionManager {
  /**
   * Creates a new instance of the EditionManager class.
   */
  constructor() {
    // TODO: Initialize the EditionManager class
  }

  /**
   * Creates a new edition.
   * @param {string} name - The name of the edition to create.
   * @param {NFTMetadata} metadata - The metadata for the edition.
   * @returns {Promise<string>} A promise that resolves to the ID of the created edition.
   */
  async create(name: string, metadata: NFTMetadata): Promise<string> {
    // Replace the following with the actual implementation to interact with the Fuel network
    console.log(`Creating edition: ${name}`);
    return "edition-id";
  }

  /**
   * Lists all available editions.
   * @returns {Promise<EditionType[]>} A promise that resolves to an array of edition objects.
   */
  async list(): Promise<EditionType[]> {
    // Replace the following with the actual implementation to interact with the Fuel network
    console.log("Listing all editions");
    return [];
  }

  /**
   * Gets the details of a specific edition.
   * @param {string} editionId - The ID of the edition to retrieve.
   * @returns {Promise<EditionType>} A promise that resolves to the edition object.
   */
  async get(editionId: string): Promise<EditionType> {
    // Replace the following with the actual implementation to interact with the Fuel network
    console.log(`Getting edition: ${editionId}`);
    return {} as EditionType;
  }

  /**
   * Mints a new token in a specific edition.
   * @param {string} editionId - The ID of the edition.
   * @param {string} recipient - The recipient address.
   * @returns {Promise<string>} A promise that resolves to the ID of the minted token.
   */
  async mint(editionId: string, recipient: string): Promise<string> {
    // Replace the following with the actual implementation to interact with the Fuel network
    console.log(
      `Minting token in edition: ${editionId} for recipient: ${recipient}`
    );
    return "token-id";
  }

  /**
   * Updates the metadata of a specific edition.
   * @param {string} editionId - The ID of the edition.
   * @param {NFTMetadata} metadata - The new metadata for the edition.
   * @returns {Promise<void>} A promise that resolves when the metadata has been updated.
   */
  async updateMetadata(
    editionId: string,
    metadata: NFTMetadata
  ): Promise<void> {
    // Replace the following with the actual implementation to interact with the Fuel network
    console.log(`Updating metadata for edition: ${editionId}`);
  }
}
